[SQUEAMISH-121] : rework on delete global services include file

+++ Domain/Entities/Renderers/Textured/CubeTextured.cpp
@@ -195,6 +195,6 @@ namespace Renderers {
 
         };
 
-        m_bytes_textcoord_size = m_texture_coord.size() * sizeof(GLfloat);
+        m_bytes_textcoord_size = (unsigned int) (m_texture_coord.size() * sizeof(GLfloat));
     }
 }
